# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Union Bank API",
    "description": "REST API for the Union Bank Management System.\n\n**API Versions**\n- `/api/v1/` \u2014 Legacy endpoints (bare response models, backward compatible)\n- `/api/v2/` \u2014 Current endpoints with standardised `ApiResponse[T]` envelope\n\nAll endpoints return JSON. Authentication uses Bearer JWT tokens.\nUse `/docs` for interactive API documentation.",
    "termsOfService": "https://union-bank.example.com/terms",
    "contact": {
      "name": "Union Bank Dev Team"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/admin/loans": {
      "get": {
        "summary": "V2 Admin List Loans",
        "description": "View all loan applications with statistics (admin only).",
        "operationId": "v2_admin_list_loans_api_v2_admin_loans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_LoanAdminStats_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/admin/loans/pending": {
      "get": {
        "summary": "V2 Admin List Pending Loans",
        "description": "View all pending loan applications (admin only).",
        "operationId": "v2_admin_list_pending_loans_api_v2_admin_loans_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_LoanOut__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/admin/loans/{loan_id}/approve": {
      "post": {
        "summary": "V2 Admin Approve Loan",
        "description": "Approve a pending loan application and disburse funds (admin only).",
        "operationId": "v2_admin_approve_loan_api_v2_admin_loans__loan_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/loans/{loan_id}/reject": {
      "post": {
        "summary": "V2 Admin Reject Loan",
        "description": "Reject a pending loan application (admin only).",
        "operationId": "v2_admin_reject_loan_api_v2_admin_loans__loan_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanRejectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/loans/all": {
      "get": {
        "summary": "V2 Admin List All Loans",
        "description": "View all loan applications (admin only).",
        "operationId": "v2_admin_list_all_loans_api_v2_admin_loans_all_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_LoanOut__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/admin/accounts": {
      "get": {
        "summary": "V2 Admin View Accounts",
        "description": "View all registered accounts with pagination (admin only).\n\nReturns X-Total-Count header for pagination-aware UIs.",
        "operationId": "v2_admin_view_accounts_api_v2_admin_accounts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_AccountListItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/accounts/search": {
      "get": {
        "summary": "V2 Admin Search Accounts",
        "description": "Search accounts by account number or name (admin only).",
        "operationId": "v2_admin_search_accounts_api_v2_admin_accounts_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Search by account number or name",
              "title": "Q"
            },
            "description": "Search by account number or name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_AccountListItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/accounts/{acc_no}/freeze": {
      "post": {
        "summary": "V2 Admin Freeze Account",
        "description": "Freeze a customer account (admin only).",
        "operationId": "v2_admin_freeze_account_api_v2_admin_accounts__acc_no__freeze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acc No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/accounts/{acc_no}/unfreeze": {
      "post": {
        "summary": "V2 Admin Unfreeze Account",
        "description": "Unfreeze a customer account (admin only).",
        "operationId": "v2_admin_unfreeze_account_api_v2_admin_accounts__acc_no__unfreeze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acc No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/accounts/{acc_no}": {
      "delete": {
        "summary": "V2 Admin Delete Account",
        "description": "Permanently delete a customer account and all its transactions (admin only).",
        "operationId": "v2_admin_delete_account_api_v2_admin_accounts__acc_no__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acc No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/statistics": {
      "get": {
        "summary": "V2 Admin Statistics",
        "description": "View bank-wide statistics (admin only).",
        "operationId": "v2_admin_statistics_api_v2_admin_statistics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_StatisticsData_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/admin/accounts": {
      "get": {
        "summary": "Admin View Accounts",
        "description": "View all registered accounts with pagination (admin only).",
        "operationId": "admin_view_accounts_api_admin_accounts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/unionbank__entrypoints__api__main__AccountListItem"
                  },
                  "title": "Response Admin View Accounts Api Admin Accounts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/accounts/search": {
      "get": {
        "summary": "Admin Search Accounts",
        "description": "Search accounts by account number or name (admin only).",
        "operationId": "admin_search_accounts_api_admin_accounts_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Search by account number or name",
              "title": "Q"
            },
            "description": "Search by account number or name"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/unionbank__entrypoints__api__main__AccountListItem"
                  },
                  "title": "Response Admin Search Accounts Api Admin Accounts Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/accounts/{acc_no}/freeze": {
      "post": {
        "summary": "Admin Freeze Account",
        "description": "Freeze a customer account (admin only).",
        "operationId": "admin_freeze_account_api_admin_accounts__acc_no__freeze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acc No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/accounts/{acc_no}/unfreeze": {
      "post": {
        "summary": "Admin Unfreeze Account",
        "description": "Unfreeze a customer account (admin only).",
        "operationId": "admin_unfreeze_account_api_admin_accounts__acc_no__unfreeze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acc No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/accounts/{acc_no}": {
      "delete": {
        "summary": "Admin Delete Account",
        "description": "Permanently delete a customer account and all its transactions (admin only).",
        "operationId": "admin_delete_account_api_admin_accounts__acc_no__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "acc_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acc No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/statistics": {
      "get": {
        "summary": "Admin Statistics",
        "description": "View bank-wide statistics (admin only).",
        "operationId": "admin_statistics_api_admin_statistics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_LoanAdminStats_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LoanAdminStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "title": "ApiResponse[LoanAdminStats]"
      },
...
      "ApiResponse_list_LoanOut__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LoanOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[LoanOut]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}